809611: Fix undefined variable in installedtab for expired
entitlement_certificate got removed in a refactor, and
was causing an undefined variable error in
installedtab.py:InstalledProductsTab.update_products()
for expired products.

But, the previous code was also ignore cases where
there could be multiple entitlements or orders, so
fix that as well by join()'ing the subscription
numbers together for the info string.