810290: use correct calculation for "Next update" time in sm-gui
There was an error when writing the "next update" time to disk where we were
adding seconds in the minutes place. For example, an update interval of ten
minutes would incorrectly bump the "next update" time out by 600 (10 * 60)
minutes, which caused sm-gui to display the wrong date in the "next update"
area. This was only being shown incorrectly in the UI; the rhsmcertd timers
were running at the correct intervals.

The above bug has been corrected. Additionally, the variable names for the
timers have been made clearer to avoid this in the future.