add a gconf setting for users who do not want to use the icon
In some cases, we do not want the rhsm nag icon to display. The prime example
of this is if a system is used on a daily basis by a user, but controlled by a
system administrator who takes care of the system's entitlements. In this case,
the user does not need to be shown the entitlement status of the machine.

In this scenario, the sysadmin or user can add a gconf entry to make the icon
not start up. Note that this is only checked at startup time; a user would need
to log out and log back in for changes to take effect.

testing note: the command to set the flag is

gconftool-2 --set "/apps/rhsm-icon/hide_icon" --type boolean "true"