820743: some multiline strings extracted into string catalog incorrectly
This is a different fix than in RHEL6.3 branch. Here we
format the strings properly, so the msgid's will be
correct for the next round of translations.

Bug here is inttool/xgettext does not understand string
concatenation of form:

msg = _("this is an example message about " +
      "how to format strings.")

Removing the plus operator works though:

msg = _("this is an example messae about "
        "how to format strings.")