820294: Let candlepin handle org/env/key validation
We don't need to verify if the user specifies a org with an environment
or with an activation key, as the user might only have one org, or a
default org, and so not need to actually supply one on the cli. Let
candlepin take care of notifying the user they need to specify one.