828042,828068: Make ja_JP's Confirm Subscription unique for firstboot.
The rhn_review_gui screen and the rshm_confirm_subs screen
have the same translated name in japanese. Which confuses
6.3 firstboot.

Most unintrusive fix was to add a space at the end of the
translation so that the screens were unique. This is a hacky
solution, but avoids having to get new translations for each
language if we were to change the string to something new.