819665: print msg if user is registered to RHN Classic on "identity" command
If a user runs the "identity" command, they previously would not recieve
messaging if they were registered to RHN Classic or not.

Instead, the user will get a notice stating they are registered to RHN Classic.
If the user is registered to both Classic and cert-based RHN, they will get a
warning message before then cert-based portion of the "identity" command runs.