817901: Show text when there are no subscriptions to show.
The All Subscriptions tab now shows a message initially telling users to
hit Update to search for subscriptions.

As soon as we have pools to display, the label disappears and the list
of subscriptions + details appears.

Similarly if you adjust filters and we no longer have anything to show,
the list + details are hidden again, and the label appears telling the
user no subscriptions match current filters.

Considered several ways to get this text displayed, but this approach
felt cleaner than anything involving forcing fake data into the
treeview, destroying it's columns, or fighting with all GUI elements
that would need to be disabled or know the rows are not valid
subscriptions.