Alternate method of opening the proxy settings.
Instead of a label where clicking anywhere will open proxy settings, use
a TextBuffer styled to look like a label, with a text tag simulating a
link that can be clicked.

Translators will need to preserve the <a></a> tags in their messages so
we can know what text should be made into the link.