Use ISO8601 date format in allsubs tab
Previously we were using the localized date
format (ala, strftime('%s')) which is lovely,
but is more or less impossible to parse, even
by the tools that claim to (for ex, strptime('%x')).

We had worked around this in a few cases, by
falling back to en_GB localized date format,
but that was pretty much just a kluge. For
bonus fun, some of the dates would only
fail to parse on double digit months. Or for
some locales, specific months that had
unicode char's the date parser didn't like.
In addition, strptime() is especially broken
on python 2.4 available on RHEL 5 and earlier.

So, enough of that. ISO8601 dates it is.

- remove locale kluge
- remove unit tests specifically for that kluge
- add test cases for widgets.DatePicker