Refactor of SubDetailsWidget and GladeWidget
Instead of passing a list of widgets to "pull"
into each subclass init, make that list a class
variable.

Also make the glade file to use a class variable.

Trying to remove odd behaviour of creating a widget, then
deleting a bunch of widgets from it (and the glade/gtk
warnings that come with that). To this end, subdetails.glade
was split into subdetails and contractsubdetails, though
they now duplicate a little bit. It is a small amount
of glade dupe compared to the previously weird code
though.