842170: replace None service level/type with "" not None
Empty or null/None service levels were being listed as
None in list --consumed. Which is confusing since
"None" (the string)[1] is a valid service level, so
it was also wrong.

Issue was "%s" % None or "blip", the % has precedence,
so the LHS was always a non empty string (because
we sub in a translated "None" string for None objects)
and therefore never False.

[1] yeah, I know...