843191: 'version' command showed wrong info with no network
If the attempt to get the status and version info from the
candlepin server failed (for example, wrong host name or
missing certs) the server type defaulted to
'subscription management service' which is not particular
correct.

Errors were also being ignored, so at least show the
error and log it now. server-type should be "Unknown"
in these cases now.