remove the explicit url search from error handling.
Katello doesn't return nice error messages like Candlepin.  So we
need to dumb down the error handling a bit. And instead of exiting
on a 404, we simply log a not supported method and return None. This
allows us to continue subscribing in the event servicelevel is not
supported by the server.

In the event the server supports service level but an exception occurs,
we exit.