845827: Split server version checkout out to avoid errors
The changes to populate server and client versions in the logs
were causing errors in cases where network errors were being
poorly handled. Part of the reason for this was that a call
to the server /status was happening very early, and there
was not error handling setup for that.

Another cause is that /status for a deleted consumer, or
invalid consumer will fail.

This splits the client version info (python-rhsm and
susbcription-manager rpm versions) and the server version
info into two seperate calls. This way we can log the client
version even in cases of no server info, and handle server
errors better, or skip in it cases where it doesn't make
sense.

Also, use mocking more liberally in some of the test
cases as above changing exposed some odd behaviour with
the version related stubs.