Add test cases for certmgr.py (and make certlib testable)
852706 needed tests cases, and there was almost
no coverage of certmgr.py

The various *Lib and *Action classes create a ActionLock()
during at class parse time (ie, before the init runs) so
we can't really mock out ActionLock. So change certlib
to pass in it's own lock to it's invocation of CertLib.

Note: Those ActionLock() invocations need to go away, or
at least, lock should be a class attribute populated with
a ActionLock(), at least that way we would have a reference to
it.