854879: Fixes for Anaconda desktop/workstation product cert installation.
Recent performance fixes added an internal cache to the product
directory object, it only looks for new certs once refresh is explicitly
called. In this case we are writing and sometimes deleting certs and
then expecting the list() call to give accurate results but the refresh
call was missed here.

Fixed an issue where we delete the now obsolete desktop cert, but delete
the workstation product ID from the product database.

Fixed an issue where any certs after the desktop cert would be ignored
because we're returning when we should be continuing to the next
iteration of the loop.