Create a TestConfig class, and use it in test setup
Replace the StubConfig with a TestConfig, which is a
disarm'ed version of the normal config class.

Also, instead of using a stringIO of a config file
string buffer, actually create a file, write it, and
let the config class read it. Should cover a bit more
code. We write that config class to a tmp dir, and
create a tree corresponding to our /etc/ tree there
as well, and setup the test config to point there.