For cert dirs, move to using a class method to get the configured path
Previously we were setting a class variable to a config
variable. This causes all sorts of weirdisms while testing,
because the class variable gets populated before any of
the test setup can run. This made mocking the path's
tedious and error prone.

Now call a static method that get's the path from the config
when needed. It would probably be even better if it were just
a regular method, but there are other class method that
need to be able to call it.