853885: Fix icon notification popup only displaying once.
Appears related to a past fix to get the icon showing up multiple times.
We were trying to re-use the status icon, but creating the popup new
every time. This would be ok but we were also trying to clean them both
up at the end of main which indicates a possible failure to cleanup each
popup that was created.

Creating a new icon each status display causes for some strange
rendering jitters in the gnome panel. go back to using a single icon for
this.

Switch to a gtk timeout rather than sleep(), so we don't block the
thread.