852328: Report Classic and Subscription Management consistently in the version and identity commands
The version command will always report server type. It will show 'Red Hat Subscription Management' if there is no systemid field, and will show 'RHN Classic and Red Hat Subscription Management' if there is.

The identity command will omit the server type unless there is a system id field. In that case it will show 'RHN Classic' if not registered with Subscription Management, and the same both string as above if registered to both.