871146: Fix proxy errors on first yum operation after registration.
Change in behavior from RHEL 5 to RHEL 6. The previous code was doing a
check like "if value:", if false it was skipped.

New code compared with "is not None", which for empty strings evaluates to
true, where as previously this would have been skipped as well. Proxy
just happens to start out as an empty string.

Result was that on repo creation, the empty string was going in and
making yum upset, but any repo update would catch it and remove it. (as
there seem to be several places where we're watching for empty strings)

Make sure we don't let empty strings slip out into the repo file by
restoring the old behavior when getting the actual list of items to go
into each repo.

Also removing a block of code in the RepoFile that looks unecessary,
after passing in section items we were then going through each item and
setting them manually. The two attempts were from different commits, the
constructor now does this for us, I think the other version was just old
code that was meant to be removed.