874749, 874804: Use the terms Attach and Remove instead of Subscibe and Unsubscribe.
For the GUI, this is a basic string change. For the CLI, the terms subscirbe and unsubscribe are command names. To resolve this, the commands were renamed to the new name and a new command was created for the old terms. This new command is marked as deprecated, and is shown in the secondary section. Both command should work.

The only other nutty change is around auto-subscribe at register. Now, either autosubscibe or autoattach  can be used.