support 1:N cert to repo mapping in product db
There are instances of multiple repositories with the same product cert.
However, the existing product db only handles 1:1 mappings of cert to repo, so
the last discovered repo always "wins", erasing data for old product mappings.

This patch converts the old style 1:1 db to a new style that supports 1:N cert
to repo mappings by using an array of strings instead of a string. It also adds
logic to the product database to deal with 1:N mappings, as well as code to
delete the product cert if the last mapping is removed.

Expected behavior:
 * run plugin, productid.js gets updated to new format
 * install package X from repo "foo", productid.js is updated, product cert for "foo" is laid down
 * install package Y from repo "bar", which has the same cert as "foo". productid.js is updated again
 * remove package Y, productid.js is updated
 * remove package X, productid.js is updated and cert for "foo" is gone