878820: Fix console error when yum.repos.d does not exist.
Error would display on console when yum is not installed (RHEV-H) when
unregistering or subscribing.

If this directory does not exist, it's a good indication that we
shouldn't bother trying to write redhat.repo. This is effectively the
exact same as the manage_repos setting in rhsm.conf, so when creating
we'll assume manage_repos=0 if the directory does not exist.