878820: Fix no such dir yum.repos.d error when yum not installed.
Error would display on console when yum is not installed (RHEV-H) when
unregistering or subscribing.

If this directory does not exist, it's a good indication that we
shouldn't bother trying to write redhat.repo. We only write in two
locations in repolib, so update both to skip writing if the dir does not
exist.

This is effectively behaving like manage_repos is turned off in by the
config.