772936: Warn the user when clock skew is detected.
All server responses are required to contain a date header. This
header is used to look for instances where the time on the local machine
is more that 6 hours off of the server. The time is arbitrary.

This should catch cases where the local machine is mis-configured. It is
possible that it will not catch when the candlepin server clock is off
since the header may come from the apache server.