Add two manifest commands to rct.
dump-manifest will export he data from the manifest into a directory so that it can be accessed via the file system.
cat-manifest will output to the screen interesting summary data about the manifest

As part of this work, the commands are broken into 2 files. One is for certificates, and one is for manifests.