907536: Don't send body if it's just ""
We default post and put's to a body of "". Restlib._request
ends up making the body of the message be the jsonified
'""' instead of just empty with 0 content length.

Candlepin doesn't have a problem with this, but old
version of headpin do. Change default body to None, so
we use that in the cases where we don't send anything
in the POST/PUT body.