920091: Reload identity cert monitor in installed tab.
The installed tab now uses identity to lookup product valid date ranges.
This exposed a problem with how we monitor both entitlement directory,
and the identity cert for changes.

This is tricky because if we hook up monitors for entitlement dirs and
identity cert, and both change (which is the case in unregister), we
don't know which is going to trigger first. As such when either happens,
we have to reload identity explicitly.