919706: Relax rhn-setup-gnome dependency.
To allow installation on older RHEL 5 and 6 versions, this relaxes the
requirement on rhn-setup-gnome. This version requirement was added to
address Bug #810399, so by removing it and then installing on some of
those older RHEL versions means this bug may surface again. It is
however an edge case so this has been deemed to be ok.

Relaxing the version requirement should have no impact on RHEL versions
with that newer rhn-setup-gnome, so the bug will not surface anywhere it
isn't currently happening regardless.