921249: Fix Unknown virt status being reported to server.
hwprobe.py use of signals is only allowed in a main application thread
now. During GUI registration and auto-attach we were needlessly
re-loading facts which triggered this code and failed as this is an
async operation.

To solve this patch just uses the facts we loaded in the main thread
when the GUI was launched. FactLib should continue to re-load facts if
none were provided.

We also make sure the facts are loaded immediately after creation,
overriding the attempt at lazy loading.