Handle missing keys when determining environment.
In Katello commit d880732149287fb8b98f37a1ba9cb7399d90eafe, the
"label" key was removed from the JSON returned when getting a list
of the environments.  Additionally, the label value was mapped to
the "name" key.  BZ 978001 addresses these issues and adds the
"displayName" key containing the actual name value to the JSON.

This commit should allow for 3 varieties of JSON:
 - Both "name" and "label" are available.
 - Only "name" is available.
 - Both "name" and "displayName" are available.

The goal is to allow the user to enter either the name or label
of the environment and have the system register properly.