994997: Fix unknown virt status in firstboot.
Second time we've seen this code cause a bug, as it is not possible to
use this signal code outside of the main thread. This time we'll try to
replace it so the issue no longer occurs by using another method for
running the command.

Popen had issues with:
https://bugzilla.redhat.com/show_bug.cgi?id=681925

The signal code was added as a result of that bug.

Instead we will now use a different method for running virt-what
entirely, which is already in use elsewhere in the facts module.