Prevent name collision over the parent variable in RHEL 5 Firstboot.
RHEL 5 Firstboot calls passInParent and passes in a reference to
the firstbootWindow class.  This value is assigned to parent in
moduleClass.  However, moduleClass extends RegisterScreen which
also defines parent and uses it when creating a gtk.MessageDialog.
Having parent set to the firstbootWindow class when creating a
MessageDialog fails as MessageDialog expects the parent parameter
to be of type gtk.Window.