Move PackageProfileLib and InstalledProductsLib
Move these to packageprofilelib.py and installedproductslib.py

Also "actionify" PackageProfileLib and InstalledProductsLib
and add action reports.

Note: a lot of the Lib/Action hierarchy isn't entirely
needed atm, but moving to getting all of the various
*Lib classes on a similar implementation.

Aside from consistency, the eventual goal is to make the
*Lib classes running a locking context manager, and
changing from the current locking which is primary
for preventing multiple processes from contenting
(using a file lock based rerentrant lock) to something
that could be used for thread locking.