identitycertlib now uses identity injection
Add getConsumerCert to Identity to get the actual
consumer cert so identitycertlib can persist it
if need be.

move the actual ConsumerIdentity.read() to a
_get_consumer_identity() to make creatin mock
identities for injection easier.

Add an injected identity to CliCommand

Replace check_registration with either:
  CliCommand.assert_should_be_registered
    or
  CliCommand.is_registered

Otherwise use CliCommand.identity for consumer
identity information.

Update tests to remove mocks and monkey patches.