Fix firstboot issues with new firstboot.
Since we now run last in firstboot, we were hitting
a case where the threads for the subscription api
calls were not finishing before we stop checking
the idle loop for events (which in this case, were
the idle source thread watcher methods).

Also handle no rhn-client-tool modules being installed,
and skip the steps that rely on them.