1017351: ignore dbus failures on show_window
If we get a dbus exception from what seems like
a working dbus session bus connection while attempting
to raise an existing subscription-manager-gui, now
ignore the error, and open another window. And log
some of the info, so we can track this down.

Getting into that scenario seems to only get triggered
from tests automation when using the at-spi framework.
This isn't a fix, but it should get tests working, and
more info.