Properly update cache when overrides are modified
When we modify the overrides on the server, the current
set of overrides are returned, so there is no need to have
the cache fetch them again when it is updated.

Added an update_local method to the cache so that the current
cache could be reset to the specified overrides.