Improve display of the system info on facts dialog.
Alignment was getting out of hand particularly with addition of the new
consumer UUID display, which is quite long.

Use a table and reserve most of the space for the values, allowing for
display of the long consumer UUID.

As before we will hide the environment and org ID if they are not
available, but we continue to show the org name as Unknown in this
situation.

Switched to showing Unknown for the System Identity as well rather than
just hiding it.