998033: Handle Unauthorized/Forbidden exceptions in CLI/GUI
- translate and print messages when caught in CLI
- translate and show messages when caught in GUI

This bug fix was required as we are now handling
AuthenticationException directly in CLI/GUI which
is the base exception for Authentication/Forbidden
exceptions, which would result in a generic message
instead of custom messages per exception type.

Exception handling in CLI/GUI could use some work
to remove the duplication, but that will be left
for another commit.