respect http(s)_proxy env variable for proxy information
The idea behind this is to allow a user to specify the username/password
and other proxy information in an environment variable instead of
the configuration file. Subscription manager will use the values passed in from
the cli or from the configuration file if present otherwise it will try to use
the http(s)_proxy environment variables. The variables are read in the following
order:

HTTPS_PROXY
https_proxy
HTTP_PROXY
http_proxy

Summary of changes:

    * moved parse_url, ServerUrlParse* exceptions, has_(good|bad)_scheme
      to python-rhsm
    * cp_provider, managergui, managercli, repolib use the new
      utility method from python-rhsm.
    * updated rhsm_login to use the new utils package from python-rhsm
    * migrate were untouched.

use rhsm.utils
bump python-rhsm requires