1033741: Refresh Overrides CP connection when dialog is shown
Since the RepoOverrides dialog is created when the subscription-manager
GUI is first created, we need to make sure that we recreate the
connection to the server b/c it may have been changed when we register.

The connection is being cached in the Overrides object, so it never
recieves the updated connection configuration if it were changed during
registration.