1050850: re-evaluate system facts when checking for updates
Checking for updates would be a noop after the first time because
we assumed that system facts would not change very quickly. Now
we check every time has_changed is called to more accurately
determine whether the facts have changed.

Added cpu speed from lscpu to the graylist because it is the
same data we're already graylisting from another source.