874169: Fix label alignment in progress UI
We were doing some funky alignments with the progress
bar widget that were causing different label alignments
as the string length changed.

Removed the glade alignment constraints on the progress
bar and the margin on the label so that everything lines
up as expected.

Previously we had a large left margin on the bar and
label that offset from the bold progess label. I also
removed this so that all 3 widgets align. This looks
a lot better IMO.