1038242: add anaconda.pid check before chroot
The yum plugin 'productid.py' has to detect if it is
running in anaconda or not, so that it will place the
productid file in the correct chroot.

Previusly, this just checked for existince of '/mnt/sysimage'
but that can exist outside of anaconda (or in the original
bug, left over by anaconda). So add a check that
'/var/run/anaconda.pid' exists as well.