1096676: Use simplejson on RHEL 5.
There is no json module in Python 2.4 on RHEL 5.

On RHEL 6 python-simplejson is available, and performs faster than the stdlib
json module, but we will continue only using it if installed (see ourjson.py)
and not require it as an explicit dependency ourselves.

On RHEL 7 python-simplejson is not available at all.