Simplify the config model a little.
OstreeConfigUpdates now has a copy of the old config,
the new config, and a map of the content->remotes
generated, instead of just the new config.

Remove OstreeConfigController, and move to letter
OstreeUpdates handle it's persistence and updating

Add a currently poorly named OstreeContent that is
responsible for finding currently entitled content
that provides ostree content.

Change OstreeCore to be a dict, since it's contents
are likely to change, and we don't need to understand
them, just not lose them.

Fix the regex that extracts the remote name from the
repo config remote section name. It was previosly
including the quote.