Don't encourage registration in yum plugin if we have ents but no identity.
In containers, we will have entitlements but no identity, we do not want to
encourage users to run register every time they use yum in this situation.

Also fixing a logic error, if you had no identity, and no ents, we were
prompting you to subscribe. (when we should be prompting you to register)