1104498: Fix productid install in anaconda
The fix for #1038242 added a check to verify the
product-id.py plugin is running within anaconda by
checking for existence of /var/run/anaconda.pid.
/var/run/anaconda.pid does not exist on RHEL5, so
that check always fails, and the plugin does not
changes it's root dir, and installs productid to
the anconda runtime instead of the sysimage.

Remove the check for anaconda.pid.