Add tox.ini with ignores for pep8 indention
Newer pep8 wants stricter indentions styles, particularly
on continuation lines. Code needs to cleaned up for that,
but this gets pep8 back to out expected checks.

E12* are all indention related.
E713,E714 are legit suggestions for 'not a in b' v 'a not in b'
  that need to be addressed.

Remove ignores for Makefile, move to tox.ini.
May break on old pep8.