1108257: special case prod tag rhel-5-workstation
The product certs for RHEL5 Client Worktstation
are a little odd, and only include the product
tags 'rhel-5-workstation,rhel-5-client-workstation' even
though it's a RHEL variant (should also include 'rhel-5')

Since rhelproduct wasn't matching that as a RHEL variant,
it wouldn't consider it as a candidate for the 'release'
command. Add a specical case for 'rhel-5-workstation'.