Force subscription-manager yum plugin to respect the managed root
Originally, the plugin updated the repo file of the host even if the Yum's
root was different. It did not make any sense. If a chroot is managed,
user expects to work with actual repos there.

To obtain Yum's root, we need to wait until it's configuration is processed.
Thus we have to hook to the following slot. As long as we modify only repo
files, it's ok.