Handle Content.gpg="http://" as gpg-verify=false
Content.gpg is a url, not a boolean. So for cases
where it needs to be both, except the "http://"
value as equilivent to disabling gpg-verify.

Flip gpg-verify to enabled by default for ostree
content, and disabled with above url.

We will set gpg-verify=true always, unless the
Content.gpg attribute exists is = "http://".
Missing, empty, None, False,etc values for
Content.gpg do not flip or unset gpg-verify.

Add spot checks of gpg-verify to various unit
tests.