make has_ostree use macro value NOT hardcoded value.
has_ostree was being set to use_systemd which caused it to always
be set and considered TRUE. We really wanted it to be the *value*
of use_systemd. In order to do that you need to use the percent
sign to make it a macro so it gets expanding during build time.