Always update ostree refspec when adding remotes.
For a fresh install the ostree origin refspec for the deployed
origin may have a name that doesn't match the remote
names for the ostree content that system has subscribed
to. So, for now, always up the refspec to map the ostree
repo configs we create, so that ostree will find them.

An example of this might be a system that has it's
ostree refspec setup by anaconda to 'ostree-atomic:ostree-atomic-awesome'.
Those names may not map to the names from the product ids installed
on the machine. And if the remote names don't match the
install product names or refspec, we have to pick something.

If there is just one remote, choose it. If there is more than
one, pick the first one sorted by name. For now.

Update unit tests.