1065158: Prompt for environment on registration when necessary
Behavior when an environment is passed via an argument is
unchanegd, as well as behavior when environments are not
supported by the server.
If an owner has zero environments, the user will not be prompted
and a null value will be passed.
If the owner has a single environment, that will be used
by default