1115499: Allow enable/disable repos in same command.
Combining was previously possible, but all enable options were processed
before disable, meaning --disable * --enable a would end up with a disabled.

Uses an optparse callback to process CLI arguments and store an ordered list
of tuples of enable/disable + repoid.

Assumes the user wants them processed in the order they appear on the command
line, the last match will be the one that wins for a particular repo.