Consolidate the register and subscribe steps.
This is necessary because when registering with a consumer ID,
subscription manager currently has a bug where it returns a non-zero
exit code even though registration occurred successfully.  Since we rely
on the exit code to tell if an error occurred or not, the migration
script aborts prematurely.  By adding a --auto-attach to the
registration step we can avoid the issue.